Complete plugin-configured provider panels (Fixes #391) (#699)
* Complete plugin-configured provider panels (Fixes #391)

* Finish provider panel host contracts (Fixes #391)

* Keep provider panel tests within source-size policy

* Normalize provider mouse test module formatting

* Harden provider panel interaction details

+++ src/domain/plugin/limits.rs
@@ -66,7 +66,7 @@ pub const SCREEN_ID_LIMIT: usize = 32;
 pub const CONFIG_FIELD_LIMIT: usize = 128;
 
 /// Lowest accepted configuration schema version.
-pub const CONFIG_SCHEMA_VERSION_MINIMUM: u32 = 1;
+pub const CONFIG_SCHEMA_VERSION_MINIMUM: u64 = 1;
 
 /// The only manifest schema this executable reads.
 pub const MANIFEST_SCHEMA: u32 = 1;
